Pull request

+++ fs/operations/operations.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"math/rand"
 	"path"
 	"path/filepath"
 	"sort"
@@ -401,6 +402,21 @@ func Copy(f fs.Fs, dst fs.Object, remote string, src fs.Object) (newDst fs.Objec
 	return newDst, err
 }
 
+// SameObject returns true if src and dst could be pointing to the
+// same object.
+func SameObject(src, dst fs.Object) bool {
+	if !SameConfig(src.Fs(), dst.Fs()) {
+		return false
+	}
+	srcPath := path.Join(src.Fs().Root(), src.Remote())
+	dstPath := path.Join(dst.Fs().Root(), dst.Remote())
+	if dst.Fs().Features().CaseInsensitive {
+		srcPath = strings.ToLower(srcPath)
+		dstPath = strings.ToLower(dstPath)
+	}
+	return srcPath == dstPath
+}
+
 // Move src object to dst or fdst if nil.  If dst is nil then it uses
 // remote as the name of the new object.
 //
@@ -423,8 +439,8 @@ func Move(fdst fs.Fs, dst fs.Object, remote string, src fs.Object) (newDst fs.Ob
 	}
 	// See if we have Move available
 	if doMove := fdst.Features().Move; doMove != nil && (SameConfig(src.Fs(), fdst) || (SameRemoteType(src.Fs(), fdst) && fdst.Features().ServerSideAcrossConfigs)) {
-		// Delete destination if it exists
-		if dst != nil {
+		// Delete destination if it exists and is not the same file as src (could be same file while seemingly different if the remote is case insensitive)
+		if dst != nil && !SameObject(src, dst) {
 			err = DeleteFile(dst)
 			if err != nil {
 				return newDst, err
@@ -960,7 +976,10 @@ func HashLister(ht hash.Type, f fs.Fs, w io.Writer) error {
 func Count(f fs.Fs) (objects int64, size int64, err error) {
 	err = ListFn(f, func(o fs.Object) {
 		atomic.AddInt64(&objects, 1)
-		atomic.AddInt64(&size, o.Size())
+		objectSize := o.Size()
+		if objectSize > 0 {
+			atomic.AddInt64(&size, objectSize)
+		}
 	})
 	return
 }
@@ -1475,6 +1494,31 @@ func moveOrCopyFile(fdst fs.Fs, fsrc fs.Fs, dstFileName string, srcFileName stri
 		return err
 	}
 
+	// Special case for changing case of a file on a case insensitive remote
+	// This will move the file to a temporary name then
+	// move it back to the intended destination. This is required
+	// to avoid issues with certain remotes and avoid file deletion.
+	if !cp && fdst.Name() == fsrc.Name() && fdst.Features().CaseInsensitive && dstFileName != srcFileName && strings.ToLower(dstFilePath) == strings.ToLower(srcFilePath) {
+		// Create random name to temporarily move file to
+		tmpObjName := dstFileName + "-rclone-move-" + random(8)
+		_, err := fdst.NewObject(tmpObjName)
+		if err != fs.ErrorObjectNotFound {
+			if err == nil {
+				return errors.New("found an already existing file with a randomly generated name. Try the operation again")
+			}
+			return errors.Wrap(err, "error while attempting to move file to a temporary location")
+		}
+		accounting.Stats.Transferring(srcFileName)
+		tmpObj, err := Op(fdst, nil, tmpObjName, srcObj)
+		if err != nil {
+			accounting.Stats.DoneTransferring(srcFileName, false)
+			return errors.Wrap(err, "error while moving file to temporary location")
+		}
+		_, err = Op(fdst, nil, dstFileName, tmpObj)
+		accounting.Stats.DoneTransferring(srcFileName, err == nil)
+		return err
+	}
+
 	if NeedTransfer(dstObj, srcObj) {
 		// If destination already exists, then we must move it into --backup-dir if required
 		if dstObj != nil && fs.Config.BackupDir != "" {
@@ -1501,6 +1545,17 @@ func moveOrCopyFile(fdst fs.Fs, fsrc fs.Fs, dstFileName string, srcFileName stri
 		defer accounting.Stats.DoneChecking(srcFileName)
 	}
 	return err
+}
+
+// random generates a pseudorandom alphanumeric string
+func random(length int) string {
+	randomOutput := make([]byte, length)
+	possibleCharacters := "123567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+	rand.Seed(time.Now().Unix())
+	for i := range randomOutput {
+		randomOutput[i] = possibleCharacters[rand.Intn(len(possibleCharacters))]
+	}
+	return string(randomOutput)
 }
 
 // MoveFile moves a single file possibly to a new name
